+++ SwiftBuddiesIOS/Targets/CoreModule/Sources/Core.swift
@@ -1,0 +1,3 @@
+class Hello {
+    
+}
